Initial Commit
I felt that with my current skill set, and the skills that would be
required by any future contributors that python is going to be the core
technology in our stack. The web application for clients to consume
should be simple and will benefit from a traditional server rendered app
and not use something heavy client-side. A monolithic, opinionated
framework is also going to be a big benefit for this project. A lot of
the initial prototyping was done with node.js and react, but I've
decided to move away from that stack because it doesn't provide and
advantage at the core of this product.