Add 'layoutstyles' block on base
This is needed so 'layout.html' can load more 'global' styles that
should be included on everything extending the layout. We don't want all
children to have to know to include these styles, but have the chance to
overwrite them if necessary. This allows the children to load the styles
they need while also not affecting the globals.