Create a new project
Users or Organizations should be able to create a new project. As of
right now only users can create projects, but project are owned by an
'Owner' model. This model allows an entity to be owned by either a User
or an Organization. There should be an Owner record each permutation of
User <-> Organization relationships.